@@ -1,18 +1,18 @@
 name: fluq
 description: A simple state management api based on streams with query and mutation handlers
-version: 0.0.2
+version: 1.0.0
 repository: https://github.com/mceck/fluq
 homepage:
 
 environment:
-  sdk: '>=2.7.0 <3.0.0'
+  sdk: '>=2.12.0 <3.0.0'
   flutter: '>=1.17.0'
 
 dependencies:
   flutter:
     sdk: flutter
-  provider: ^5.0.0
-  rxdart: ^0.26.0
+  provider: ^6.0.3
+  rxdart: ^0.27.4
 
 dev_dependencies:
   flutter_test:
