@@ -3,18 +3,18 @@ import 'package:flutter/material.dart';
 import 'package:fluq/fluq.dart';
 
 /// Widget for handling mutations
-class Mutation<T> extends StatefulWidget {
-  final Future<T> Function() fetch;
-  final void Function(T result) onSuccess;
-  final void Function(dynamic err) onError;
-  final void Function(Fluq cache, T result) update;
+class Mutation<ResultType> extends StatefulWidget {
+  final Future<ResultType> Function(Fluq cache) fetch;
+  final void Function(ResultType result)? onSuccess;
+  final void Function(dynamic err)? onError;
+  final void Function(Fluq cache, ResultType result)? update;
   final Widget Function(BuildContext context, Future Function() fetch) builder;
   final bool autoFetch;
 
   const Mutation({
-    Key key,
-    @required this.fetch,
-    @required this.builder,
+    Key? key,
+    required this.fetch,
+    required this.builder,
     this.update,
     this.onSuccess,
     this.onError,
@@ -22,10 +22,10 @@ class Mutation<T> extends StatefulWidget {
   }) : super(key: key);
 
   @override
-  _MutationState<T> createState() => _MutationState<T>();
+  _MutationState<ResultType> createState() => _MutationState<ResultType>();
 }
 
-class _MutationState<T> extends State<Mutation<T>> {
+class _MutationState<ResultType> extends State<Mutation<ResultType>> {
   @override
   void initState() {
     if (widget.autoFetch) _fetch(context);
@@ -33,16 +33,17 @@ class _MutationState<T> extends State<Mutation<T>> {
   }
 
   Future _fetch(BuildContext context) {
-    return widget.fetch().then((result) {
+    final fluq = Fluq.of(context);
+    return widget.fetch(fluq).then((result) {
       if (widget.onSuccess != null) {
-        widget.onSuccess(result);
+        widget.onSuccess!(result);
       }
       if (widget.update != null) {
-        widget.update(Fluq.of(context), result);
+        widget.update!(fluq, result);
       }
     }).catchError((err) {
       if (widget.onError != null) {
-        widget.onError(err);
+        widget.onError!(err);
       }
     });
   }
